+++ docs/.vuepress/config.js
@@ -33,6 +33,7 @@ module.exports = {
                 text: 'How It Works',
                 items: [
                     {text: 'How to increment the counter', link: '/how-to-increment-the-counter/'},
+                    {text: 'How to decrement the counter', link: '/how-to-decrement-the-counter/'},
                     {text: 'Using a unique key', link: '/using-a-unique-key/'},
                     {text: 'Sort by most viewed', link: '/sort-by-most-viewed/'},
                     {text: 'Get view totals per Element', link: '/get-view-totals-per-element/'},
@@ -54,6 +55,7 @@ module.exports = {
             '/': [
                 'getting-started',
                 'how-to-increment-the-counter',
+                'how-to-decrement-the-counter',
                 'using-a-unique-key',
                 'sort-by-most-viewed',
                 'get-view-totals-per-element',